针对有ttl定义的，则放入缓存中加入ttl标记

+++ src/main/java/corner/cache/annotations/Cacheable.java
@@ -75,5 +75,5 @@ public @interface Cacheable {
 	 * @return
 	 * @since 3.1
 	 */
-	public long ttl() default -1;
+	public int ttl() default -1;
 }
